Rework plot and contour plots
The existing code has the "get_xxx_plot" routines call the
_prepare_plotobj method, and the "plot_xxx" calls _plot,
which can then call _prepare_plotobj. Or at least that's the
idea, sometimes it gets more complicated (particularly
in the astro version, handing PHA data).

The contour plots also call _prepare_plotobj but then use the
_contour method. The following changes also apply to the
contour code.

This commit changes the plot code so that the get_xxx_plot
methods now directly call the prepare method of the object,
and then the plot version can call the get_xxx_plot version
(rather than indirectly calling this logic), and the plotting
is now done by a _plot2 method (which is essentially the
"plot" part of the _plot method).

This now makes it more explicit how and when the plot objects
are created, and the data that they need. This makes it easier
and cleaner for the astro layer to add new logic. It also
avoids (some) repeated code in the plot methods.

In the astro layer there is now use of the superclass for
handling non-PHA datasets (rather than replicating the
functionality of the superclass).

As part of these changes some of the keyword arguments have
now been made explicit, where they had just been "**kwargs".

A number of routines re-implemented the plotting code (e.g.
setting up the plot, drawing it, then cleaning up). These
have now been converted to calling the _plot2 helper.

Note that existing code that is now unused has not been
removed in this commit.