Refactor: remove _get_model/source methods
It took me some time to track down code changes between the sherpa.ui
and sherpa.astro.ui modules which were down to the fact that some
called _get_source and some get_source, but they are the same
function (in reality). I've therefore dropped the _get_source and
_get_model versions as we have enough difficulty tracking things
down without adding more potential confusion.

Since the removed functions begin with _ it's a Pythom idiom that
they are considered internal methods, so it's safe(ish) for us to
remove them (people should not be relying on them, and note that
the _get_source/model methods had no documentation).