Rework the int/reg error analysis plots
The get_int_unc method now takes a fast argument when it didn't
before. There have also been some changes to the replot/recalc
settings which may have improved the behavior (it looks like it
may not have been used in some cases but it's not 100% clear
from the code/my understanding).