plot/contour commands now support multiple values per command
We now grab all values that are not an allowed command and use that
in the get_xxx_plot/contour command.

At the moment this is only useful for background plots, where we
can now specify background components other than the default value
(1).

As an example:

  plot('bkg', 'bkgmodel', 'bkg', 1, 2, 'bkgmodel', 1, 2)

displays the background and model of bkg_id=1 and then bkg_id=2.