get_data/model_plot_prefs now takes an id (fix #672)
The get_model_plot_prefs() return value now depends on the dataset
identifier (e.g. DataPHA, Data1DInt, or everything else). The
get_data_plot_prefs routine also accepts an identifier, but does
not use it.

Note that this is not a perfect solution to #672, since it doesn't
support controlling the model display in a plot_fit call (as this is
handled differently to plot_model), but it's not obvious how to
handle the strange asymmetry in our plotting support of PHA files.