parse_expr now errors out if given a:b:c
The code had a check for parsing the 'a:b' token, reporting an error
if it did not match, but the logic was wrong so it could never be
reached. We now explicitly error out if there's more than one
separator (':') in a range.

There may be existing code that, for whatever reason, assumes you
can contain extra (unused) information, so this change may need to
be reverted.