Fix subtle create_expr/DataPHA.get_filter issue
The sherpa.utils.create_expr has been reworked to be simpler,
which fixes a minor issue: output was '1,2,...' rather than
'1-2,...' for certain conditions. For discrete datasets it should
make no difference, and even for integrated datasets it shouldn't
matter, but is a small improvement. The main change to the routine
is that it now errors out if it is called with a mis-matched
value and mask arguments: this is when the number of True items
in the mask does not match the length of the value array.

The sherpa.astro.data.DataPHA.get_filter routine would trigger
this error when group=False and the dataset was grouped: the mask
array was set to the grouped mask (self.mask) rather than the
ungrouped mask (self.get_mask()).

Fortunately there's surprisingly little breakage by this bug,
as most tests that are changed are ones we've just introduced to
identify the problem. The main case that a user would notice
the difference would be

    from sherpa.astro import ui
    ui.load_pha('sherpa-test-data/sherpatest/3c273.pi')  # a grouped dataset
    ui.notice(0.5, 7)
    ui.ignore(2, 3)
    ui.set_source(ui.powlaw1d.pl)
    ui.plot_model()

Before this fix, the model is plotted over the approximately
0.5 to 7 keV range (the actual limits are more permissive due
to the relatively wide bins, particular at high energy),
including the 2-3 keV range. After the fix the excluded range
is no-longer included in the model display.

You also see this problem (with the above setup):

    d = ui.get_data()
    d.get_filter(group=True, format='%.4f')
    '0.5183:1.9199,3.2339:8.2198'
    d.get_filter(group=False, format='%.4f')
    '0.4745:0.7811,0.7957:9.8623'

The group=True version is sensible (it's roughly 0.5:2, 3:7),
but the group=False is wrong.