Move channel<->group mapping to DataPHA._from/to_channel
Rather than have random parts of code deal with convertng between
channel and group number, move this logic into the _to/from_channel
routines of DataPHA (when units=channel).

I wonder if we should remove "group" numbering entirely, in which
case the _to/_from_channel routines can go back to being the
identity function. However, we will still need some code to
handle the conversion, so maybe this is as good a place for it
as we can get.

There's a few little changes of code that we can either now make,
or could have made beforehand.