FIX filtering in channel space of grouped data (fix #288, #416)
The notice method (which ignore calls) of the DataPHA class has
special handling of channel units when the data is grouped,
because we want users to continue using channel numbers, but
we have to convert into "group" numbers. The logic for this was
complicated, and seemed to enforce behavior (e.g. pick one-less
than the actual group for the upper limit in certain situations)
which do not seem warranted. With a much simpler check - i.e.
just use the actual groups - we now see the same behavior of
filtering and grouping with analysis=channel as we do with
analysis=energy or wavelength.

An example of the change in behavior, using the 3c273.pi PHA
file in our Sherpa test suite, which checks that setting the
grouping to a "width" of 50 channels per group, returns the
same data whatever the analysis setting:

    expected = np.array([213, 136,  79,  47,  47,  29,  27, 18])
    for analysis in ['energy', 'wavelength', 'channel']:
        pha = sherpa.astro.io.read_pha('3c273.pi')
        pha.ignore(None, 1)
        pha.ignore(7, None)
        pha.set_analysis(analysis)
        pha.group_width(50)
        dep = pha.get_dep(filter=True)
        assert pha == pytest.approx(expected)

This now works (and is test_grouping_filter in
sherpa/astro/tests/test_astro_data.py) but before this commit
the channel test failed since it returned expected[1:] (i.e.
had dropped the first bin). This case was because of the broken
mapping in notice() from channel to group number (as the
group_width call ends up calling notice after applying the new
grouping scheme).

Also of note is that ignore/notice no-longer error out if,
when with analysis=channel, the lower limit is set to a value
larger than the data.

Several tests are changed (either to remove an xfail decorator
or to change the value that is being searched for).

As part of this work issues #288 and #416 are fixed.

It is not immediately clear to me that we can not drop the
"group" numbering scheme, but it would be an invasive change
(e.g. the _channel_to_energy and inverse would need to change),
so it's left for a rainy day.