Tests: change test_astro_sim.py::test_pragbayes_pcaarf_limits
The test_pragbayes_pcaarf_limits test uses a random seed to
run a MCMC chain which has been engineered to trigger a number
of "problem" cases (e.g. values exceeding the parameter limits).

Unfortunately this test occasionally fails, so I've selected a
random seed to run this test with which works on my machine.

A new fixture has been added - reset_seed - which can be used for
tests for which we want to force a random seed. Note that it
resets the seed to a "random" value after the test as I want to
make sure we only force it when required.