DataPHA data is now drawn as a histogram
So you can now say for a PHA dataset

  plot_data(linestyle='-', marker='')
  plot_bkg(linestyle='-', marker='', overplot=True)

and see the histogram nature of the data.

The plot classes remain somewhat messed up, and could benefit from
a rework.

The error analysis for the data plots has been slightly-abstracted,
but there's still several variants (probably for good reason but it's
not clear).