xspec: integrate setting of XSPEC models is now fixed.
The integrate setting now matches the model type - so True for
Additive and False for Mutiplicative - and can not be changed.

However, to support old code that may have tried to change the
setting [*], we allow a call to try and change the integrate flag,
but do not actually change it. There is currently no warning about
this. Should there be (I can't make up my mind).

The current meaning of the integrate flag is not well defined, so
this is a start to providing some meaning.

[*] Any save_all script created prior to this PR by Sherpa which
    contains any XSPEC multiplicative models will try this