Histogram-style plots can now access the x attribute
This is intended to support people who have been using the UI level
get_data_plot() and get_model_plot() functions to access the data
from a plot of DataPHA or Data1DInt data. The code will be
littered with direct access to the x attribute, since they used
to be a "Plot" class. Although using xlo and xhi si likely better,
it seems like it's not worth breaking code that worked (modulo some
issues with the first and last bin).

There is the possinility that existing code using these structures
was accidentally doing the wrong thing, in which case breaking
things might be beneficial, but I doubt our community would
appreciate this.