Tests: switch to a fixed seed for sample tests
The idea is that now we have the reset_seed fixture it's easy to
set a random seed for a test. This means we can remove the lax
testing on a number of parameters (not all test cases have been
changed but this should catch those tests I've seen that are
problematic).

It is strange but a number of these tests need relatively large
tolerances (e.g. rel=1e-3) since the values seem to vary slightly
on OS and Python and/or NumPy version (even when things like the
statistic value can use the default pytest.approx tolerance).