FIX handling of the default background id for PHA datasets
Until now the default background id was set to match the default
dataset id, but this is problematic since there is no way to
pass this identifier to the DataPHA class when creating the
data objects, and so there is a mis-match between what the sherpa.astro.ui
level and the DataPHA object thinks is the default background identifier.

We now have a separate _fix_background_id method in the sherpa.astro.ui
Session, to make it explicit what is being checked, and this routine
will return the DataPHA.default_background_id setting now.

We could pass an identifier for the default background dataset to
the DataPHA clas, but that is left for another time.