Improve support for the binary-op regrid method
Initial attempts, now long lost to the git archive, tried to use
multiple dispatch to identify what method should be used to
regrid a binary pair. Unfortunately it didn't work well, since
the decision needs to be made on the whole tree rather than just
the top-level pair of models that are being combined.

Rather than use multiple dispatch to identify what regrid method to
use when combining models, use the integrate setting instead. This
requires searching through the hierarchy - such as
   mdl1 * mdl2 + mdl3 * (mdl4 + mdl5)
and then having to combine settings

  left   right  -> combined
  False  False     False
  otherwise        True

This is basically saying that multiplicative combined with multiplicative
remains multilicative, whereas any other combination is moved to
additive. The choice is left-biased, which is potentially important
because the chosen regrid method is taken from the class (and not the
instance method), so it could be changed if we add classes with
subtly-different regrid semantics. As we do with XSPEC multiplicative
models. However, it should only be an issue when combining multiplicative
style models only (i.e. those with integrate=False), and it is not
obvious if we are realistically going to come across a model we want
to regrid that contains only integrate=False components.

I am not particularly happy this approach, but I haven't been able to
come up with a better solution.