Unary/Binary models copy the integrate setting, if set
The integrate setting is copied from the components of a *OpModel
when it makes sense (they all exist and have the same value).
However, thanks to the design (using multiple parent classes and
that one of them is ArithmeticMidel) means that when there is
no setting, we end up with an integrate setting of True.

To be honest, the integrate setting of a composite model is hard
to identify when there are multiple components and they disagree,
such as XSphabs() * XSapec(), where a multiplicative
(so integrate=False) model is combined with an additive (so
integrate=True) model

- do we set it to True
- do we set it to None
- do we record all the components settings

Setting it to None has an appeal, but then how do we use that
information (e.g. downstream when deciding twhat to do with a
regrid call)? Although, to be honest, setting it to True may
also not be useful.