Unary/Binary models copy the integrate setting, if set
The integrate setting is copied from the components of a *OpModel
when it makes sense. For binary models, if both components are
False then the result is False, otherwise it's True. It is not
entirely clear, but for most cases it should default to True if
not explictily set (thanks to using multiple parent classes and
that one of them is ArithmeticModel) but it's not entirely clear
to me (e.g. ArithmeticFunctionModel should have no integrate
setting).

The idea is that, when combining models, we are going to want
to treat it like an additive model *unless* we know that both
sides are not additive.