XSPEC table models: support regrid
This uses explicit XSPEC classes rather than the generic support
that was used before.

Of note is that we address an odd issue noted with the multiplicative
table model - we can now improve the tolerance and we no longer
have a strange-appearing warning about evaluating the model with
a single grid.