XSPEC convolution models: support regrid
Switches the convolution models to use XSPEC classes rather than the
generic classes. It's not obvious this is a net benefit for the
multiplicative case, but it is a strange case (my current thought is
that convolution models realy require an additive term rather than
multiplicative, but that's a hunch). There's also a potential issue
since XSConvolutionModel has a regrid, but XSConvolutionKernel does
not.