Move the regrid code out into a separate method
The idea is to make it easier for sub-classes to access the regrid
code, but all this does is move the logic out into a separate
routine *and* cleans up the logic a little bit. The docstrings
have been expanded and an attempt to use consistent parameter
names has been made (to distinguish between a data "space" and
just giving the actual limits).