UnaryOpModel: add regrid support
The regrid method is supported if the input model supports a regrid
method. One interesting thing here is that BinaryOpModel already
extended RegriddableModel (which is just a thin wrapper around
ArithmeticModel) so this just updates UnaryOpModel to match.

The method for finding a regrid method to use is slightly-different
in the two cases (unary vs binary).