Add plotting fixture to support switching backends
The override_plot_backend fixture will run the test with all
available backends - that is the dummy backend and, if available,
the pylab backend - restoring the original choice at the end.

Note that the fixture isn't currently guaranteed to work with
the ui modules or the Session classes: they are untested.

There are no tests that use this fixture (they will be added in
a later commit).