Add notebook support for Astronomy data objects
The display depends on the backend - is it just a textual display
or do we show a SVG. Unlike the basic data objects this code,
as it does more, is not as clean for backend-specific code: that
is it assumes pylab-specific functionality. This will need reworking
when we get another plot-capable backend.