Docs: include the example notebook in the RTD pages
Fortunately including a Python notebook is easy, but there's some
choice of how to handle this. As I am including the notebook from
outside the docs/ directory I have decided to copy the .ipynb file
at run time, rather than creating a symbolic link, or having two
copies.

The notebooks/ directory contains a nbmapping.dat which indicates
where the files get copied to in the docs/ directory: it is a
tab-separated pair of notebook name and documentation location.

A new section of the main page ("Notebooks") has been added and
the notebook placed there. The decision of whether to have a
separate section like this or place the notebooks within sections
can be made once we make more notebooks.