Add tests showing issue #959
I add several tests of the cache code, using low-level tests
that the digest is calculated correctly (although as we just
repeat the code from the sherpa.models.model.py::modelCacher1d
routine there is no external validation).

There are two tests that show #959 - the original problem and
one that I used to identify the problem. It is worth keeping
both as they do test slightly-different scenarios and are not
expensive tests to run.