Add sherpa.citation to get the citation information from Zenodo
We have a chicken-and-egg issue: we need to have created the release
to get the citation information, but we'd like to include the
citation information in the documentation. This could be done with
some careful work with our release process, but we can also just
brute-force it by retrieving the information from Zenodo when needed.

The citation function defaults to version='latest' which will
query Zenodo. If a specific version is given, such as version='4.12.1',
then a table of known releases is queried, and if found, used to
create the output without the need to query Zenodo. If the version
is not known then we query Zenodo and loop through all the Sherpa
records. This is to support the case when a new release has been
made but the hardcoded table has not yet been updated (in most
cases this will be the same as using version='latest'). I have
checked that this works by temporarily removing an entry from the
hard-coded list.

If there's no match from Zenodo then we return a "default"
citation (which is also added to the end of any Zenodo citation).

This is not ideal for scripts.

MERGE

MERGE

Blargh

tmp_path is not supported in old pytest versions, such as 3.8.1
which we are using in Python 3.5 tests.