Docs: avoid invalid escape sequence warning
The main change here is to replace \sigma and \pi by sigma and pi
since the text should either be \\sigma \pi or we change this to
a r-string. As the text isnt' actually LaTeX it's safest just to
remove the \ characters.

There are a few changes thanks to adding a space, or four.

Several lines have trailing whitespace removed, so the diff looks
odd for them (ie no obvious change).