Change to reporting the "current" citation version by default.
Following a suggestion by @pkgw this now looks for the current
release by default. This means that users who have a pre-release
version will get either "old" information (a release that does not
reflect the code) or, close to a new release, a note about there
being no match. As this is not really intended for development
code I think we can live with this.