Fix issue #977 for PHA data and model plots
The code now checks for "close" edges and replaces them so that
they are identical. In this case the sherpa.utils.sao_fcmp method,
with a tolerance of numpy.finfo(numpy.float32).eps, is used to
identify close values. This works for the problematic case found
in #977 but is not a guarantee for all cases. if the difference is
much larger than I argue that the PHA or response file is "not
ideal" and really it's an issue for the data producer.

I could not see any obvious code in sherpa.astro.data or
sherpa.astro.instrument that worries about such differences. This
should really be addressed but is a much-larger issue than this
fix.

One possibility for fixing #977 was to place the logic into the
sherpa.plot histogram code, so that all histogram plots would
benefit. I did not do this because for the PHA data we have
domain knowledge that the bins are not going to relevant at
very-small values (e.g. float32 eps sizes), whereas in the
general case we do not. It therefore seems safer to only
apply a fix for these cases.