Tests: add tests showing #981
These tests require matplotlib (since you only see the problem when
you create the plot, not with the data returned by get_data_plot).

Tests that fail are marked as xfail. Unfortunately, thanks to the
way the tests are structured, several tests are marked as xfail even
though they pass, but I think that's okay here.