PHA data plots now honor the set_xlog/ylog setting
The trick is to add an entry into the session's _plot_types
dictionary. An attempt has been made to reduce the hard-coded
nature of the astroXXX checks when accessing this dictionary.

A subtle bug with the clean method and DataPHA plots has also
been addressed, and the test_datapha_plot_after_clean written
to test this out (it would have failed without the changes to
the sherpa.astro.ui.utils.Session.clean method).