Tests: rework so tests work without a plot backend
It turns out we can run the set_xlog/xlinear/... tests without
a plot backend since the xlog/ylog settings are set in the plot
object.

It makes the code a little more complex as we have to check against
the internals of the plot objects (i.e. switch between plot_prefs
and histo_prefs), but it means we don't need matplotlib support.