Add 4.12.2 to hardcoded list and add make_zenodo_release.py script
The scripts/make_zenodo_release.py script makes it easy to add a new
release to the sherpa._get_citation_hardcoded routine. I have used
this to re-create the list for all releases (just to ensure a
consistent system), which also happens to fix DougBurke to
'Doug Burke'.

I've added an entry to the developer documentation describing this
process.