Tests: allow test_ui tests to be run with pytest-xdist
The pytest-xdist runner requires that the sequence sent to
the pytest parametrize function be repeatable (so that the same
list is used when generating the tests with multiple processes).
It turns out that set doesn't meet these requirements, so change
from a set to a sorted list.

This does not change the tests that are run, just their potential
ordering, and allows you to say

    pytest -n 6

(or some other value) if you have installed pytest-xdist.