Summary ======= Fix an error in the HTML display of a model (used in the notebook) when two model components have the same name.
Details
-------
The previous implementation interated over paramters and the modelname of
the parameter can be ambiguous. Instead, we now iterate over models and then
other paramters within eahc model. That way, we can start a new block in the
html table for each new model, without relying on detection of a name change.

fixes #1013