utils: add get/plot model/source_components
Add specialised routines that will display models for the
individual components of a source (or model) expression. I
had thought to make this conditional on the model argument to
plot_model/source_component but it turns out to be surprisingly
awkward, due to how we have set up these routines. I therefore
created separate routines, which is actually a little-bit
nicer.

One question is the name for the "get" routine

   get_xxx_components_plot
   get_xxx_component_plots
   get_xxx_components_plots

I chose the middle option since it matches the other get/plot_yyy
calls.