Improve sherpa.utils.get_fwhm
We now extend down from both sides of the peak and use the
(potentially asymetric) FWHM that is calculated. If one edge does
not fall below the peak enough then the HWHM from the other side
is used (multiplied by two). If neither is found then half the
width of the X array is used, as a heuristic.