Update lev3fft test to use PHYSICAL not WCS
It is not clear why we use WCS for fitting, since it is known that
WCS fitting isn't great, particularly when you have a convolution
model like the PSF components used here.

This is being done to make some work on FWHM-related guess code
easier (since without this change the guess in the updated code will
return a FWHM in degrees (i.e. small) which messes up the fit).

We need sherpa-test-data/#24 merged before we can accept this.