Store the opstr of model combinations (unary and binary)
We store the string used to represent the operation as well as the
opreator now. Basic tests have been added to the existing operator
tests.

There is no functionality that uses this feature yet. I have found
it useful before, and have work in progress and planned that uses
this.

Note that the test has been updated to use a f-string which means
that this requires Python 3.6. This update is not required, but
we have to drop Python 3.5 at some point!