Tests: add tests of set_datadir
We check that set_datadir fails for various cases. In writing these
tests I found that the error case of "not a directory" was not
well handled (you got a NotADirectoryError with a low-level error
code rather than the OSError for other cases. This has been addressed.

This requires a bump of the minimum pytest version to 3.9, to
take advantage of the tmp_path fixture. This is an old release,
so it should be fine. It would allow other tests I recently added
to move from the tmpdir fixture, but that's out of scope here.