Tests: ensure optional id argument is tested
Either add a test including the optional id argument, or a
test without the optional id argument, for a bunch of load_xxx
routines. This is guided by covreage checks but it does not
cover all cases.

Several of the checks also actually check the load_xxx function,
as a number of the tests just seem to check we can load the data
and make no effort to check the data is as expected. There may be
other tests that check these things, but they are likely to be
more-complex tests and don't provide simple tests.