Update the existing action to use setup-miniconda
As well as moving over to setup-miniconda, the macOS setup to
install XQuatz and the macOS SDK have been moved into a single
step.

Support for the docs-only build was removed as we are not
testing the docs build on the main CI anymore.

The conda configuration is now set up in

  etc/environment-linux|macos.yml

where the "type" is set by the new matrix.ostype setting.

Note that we try to use conda to install as much as possible
as there are strange failures if pip is used to install the
pytest machinery.

**NOTE** I made an attempt to update the build/test infrastructure
but this was spectacularly unsuccessful, so I have removed
this part of the PR.

Note that test=submodule install-type=install appears to re-build
the code when we get to runnig the test suite. This is not
ideal! I have been unable to improve on this (see preceeding
note).