Allow two optimisation tests to pass
This fixes what appears to be a cut-n-paste error in the test
setup code, where the wrong size was being set for the "vector"
array. The old value was too large, which means that the previuos
version could sum up (or do something with) uninitialised
elements of this vector (beyond the values set by the methods).