Change the best-fit value of the PenaltyI test
From a review of https://www.osti.gov/biblio/6650344 I found that the
PenaltyI initializer was using the wrong npar=4 best-fit value.

The tests still pass since they have a combination of rel=1e-3 and
abs=1e-3 and I assume it's the abs=1e-3 check causing the change
here to be un-noticed.