WIP move to a more-extensible system for plot types
Rather than having the per-plot objects at the top level of the
session class, which makes it hard to identify what plots are
used in what places, and hard to write somewhat-generic code.
There was a partial attempt at improving this in #906 and #931
with the _plot_types dictionary, and this attempts to take
that idea and apply it fully:

 - the _plot_store dictionary is used to define the "names" of
   the plots (e.g. "plot", "model", ...) but it is not yet clear
   about composite plots like "fit"
 - each entry is a pair, where the first entry is the default plot
   object and the second a dictionary which maps from more-specialised
   type to plot object. An isinstance check is used to identify the
   "most specialised" type that is appropriate, relying on the
   mro depth of a class to select the "most specialised"
   I should probably use something more-specialied that a pair.

This has only been implemented for the data plot type, and so some
code is repeated (in particular _set_plot_item) that can be cleaned
up once everything has moved over.

The "old" plot objects are left in for now but will be removed.