Ensure that min/max are lists, not tuples (fix #1093)
The problem is that if the user has set min=(0, 0) but one of the
two parameters has a soft min value above 0 then the code tries
to change self.min[i], but self.min here is a tuple and so it
errors out.

We could require lists as inputs, but the existing documentation
uses tuples.

The chosen solution is to always set the min and max attributes
to a list. This has down-stream consequences

 - a consumer may be surprised that what used to be a tuple is
   now a list
 - the str output for these RegionXXX objects can change from
   a tuple to a list (shown here in a notebook test)

I believe these are acceptable changes.

An alternative would have been to something fancy to ensure the
attributes remain as tuples (just different tuples to what they
used to be). However, that is excessive complexity here.