Tests: add a test to show issue #457
Issue #457 points out that sample_flux excludes rows where a
parameter is at the parameter bounds, which - for situations
like the added test - will lead to biased results (in this
case the nh=0 values get excluded from the sample distribution).

Here we add a test to show the current behavior.