sample_flux: note that we call the sampler with clip=hard
PR #866 added parameter sampling support to the samplers but
sample_flux was not updated to explicitly support this. We still
don't provide the user access to this functionality, but we now
are explicit about what clip method is used (the default value of
'hard'), as it will make future changes clearer.