Doc improvements for sample_flux
This addresses the documentation error over the return value of
the flux values not being the first and third quartiles. The
replacement text is wordy, and it would be nice if it were easier
to read, but I'm going for "correct" here.

I have also fixed an issue in the discussion of the vals return
value, which has a shape of niter, npar + 3 (it was documented
as npar + 2). This was introduced in #866 (I think, haven't
checked the genealogy of this buglet) since when that code was
written I'd hoped that this code would have been quickly reviewed
too.