calc_sample_flux: move code out of try/finally block
There's no need for quite so much code to be within the try/finally
block, so move it. A sanity check is also added to ensure that the
samples array has the correct size (handled as an assert rather than
an actual error as it is expected that the code is used correctly).