sample_flux: use clip=soft from sample_energy_flux
Remove the clipping logic from sample_flux and rely on
sample_energy_flux.

This is a change, since we now will use the results from values
that were outside the soft limits, as they'll have been changed
to be at the limit. This will primarily change the reported
limits but could shift the median too.