sample_flux: return a statistic value even for clipped rows (fix #751)
We now return the statistic value for each row, even if it is then
clipped out of the flux calculation. Note that the statistic value
is calculated using the soft parameter limits, so the rows that
are "clipped" by this routine have a potentially confusing
mapping (since the parameter value doesn't match the value used
to calculate the statistic).

This fixes #751.