calc_sample_flux: calculate stats when outside soft limits
By using the thawedpars attribute of the model we can avoid the
soft-limit check, so simplifying the code *and* ensuring that those
rows where at least one parameter is outside the soft limit have
the "correct" statistic value (that is, they are calculated using
the parameter values).

This takes advantage of a under-documented (but I am pretty sure
it is intended, to avoid the problem we have here) feature of the
thawedpars attribute in that the soft limits are ignored when
setting the parameter values.