Ensure plot classes are exported
We domn't have a policy on what symbols we add to __all__, but it
seems to make sense to export all-but-internal classes. The ordering
in __all__ is changed to match that used in the file, as this makes
it easier to add and move them (in previous PRs the existing
ordering made it hard to handle conflicting PRs).