Move fit plots to _plot_store
Similar to the soource plot type, the handling of PHA data for the
fit plot type is complicated. In this case it is that a different
plot class is used for the model plot (to ensure the model is
grouped) compared to the model plot.

There is a subtle difference in behavior in get_fit_plot for PHA
datasets when recalc=False; we no-longer re-generate the plot
contents which better matches the super-class and the expectation
of the documentation.