Move compsource/compmodel plots to _plot_store
There is a change in behavior for plotting template models, but I
believe that the previous code did not handle code "sensibly", in
that it would only happen for plot_model_component_plot when
used from the sherpa.ui.utils layer (and possibly non-PHA data
for the sherpa.astro.ui.utils layer).

This change suggests that the new approach is not ideal, as
we need to add extra code to handle the specific PHA case, but
then again this is the place to add this sort of type-specific
check.