Move order/energy/photon/arf plots to _plot_store (astro)
Note that as entries have been removed from the _plot_types
dictionary (which will be removed at the end of this work) then
the checks for "invalid identifiers" change because of this
commit. We do not have any tests that exercise these symbols,
and the plan is that the checks will be returned once we can
use _plot_store instead.