Tests: explicit check that keywords are reserved
There are a number of plot types that are not rejected. It's
unclear whether they should be or not, since the main restriction
- I believe - is to make parsing

plot(ptype [, optional arguments], ptype [, optional arguments], ...)

easy (in that ptype cannot appear in the optional arguments section),
and these "un-rejected" plots don't really make sense for the
plot call, as they require more arguments than we expect to
pass in to a plot call.