What identifiers do we exclude?
We have several different ways to decide what sets of identifiers should
not be considered valid. Let's move to the _plot/contour_type_names
dictionary, which is used by the _multi_plot method to decide what
plots are valid, and so is set of names which is problematic.

This does allow several new names ('energy', 'lr', 'photon'). Do we want
to allow users to say

    plot('energy', lo=0.5, hi=7, num=1000)

to match

    plot_energy_flux(lo=0.5, hi=7, num=1000)

Several issues with this idea:

 - plot name is hardly descriptive (e.g. 'lr'?)
 - we couldn't support named arguments and allow multiple plots