Use a class rather than tuple + class
As we have a small set of operations we want for the "object store"
logic used in _plot_store we can go from a home-grown solution of
two-element tuple of default value and the OrderedByMRO class to
a class.

There are a surprising number of tests that are poking around at
the insides of the Session class. This is somewhat unavoidable
but I wonder how many of them we can change to using the
get_*_plot(recalc=False) method?