Store UI contour plots in a dictionary, not directly
This only affects users who are poking around in the internals of
the session class, and it's all changes to "private" attributes
so it is okay to change this.

The idea is to store the individual contour-plot objects in a
dictionary rather than a list of attributes in the session class.
This makes it easier to track from the contour data command to
the object that is used by the command, and potentially makes it
easier to extend in sub-classes of the Session class (although
this is not currently used). This is a simpler version of an
approach I plan to apply to the plot objects.