Read RMFs where N_CHAN is an array
The old code checked the number of groups, and if that is 1
assumed that N_CHAN would be a scalar. However, the OGIP standard
also allows N_CHAN to the a fixed or variable length array.
In particular, this will happen if some rows have 1 channel
subset, but others have more than one.