Add some grating related keywords for _repr_html_ for ARF, RMF, PHA
Summary
=======
Add akywords to the default output in the _repr_html_ for some X-ray classes.

Details
=======
This only changes the output of those fiels that actually have the keywords
set. This the list contains a few keywords that e.g. CIAO does not set by
default, but that could be set in grating data. For grating data, the
diffraction order and the grating type (HETG / LETG) are very important
and should be printed.