Tests: tweak the smoke tests
The location of the FITS file is only used in one test, so only
calculate it once.

Add some basic I/O tests to ensure that the file that was written
out can be read back in (it extends the test beyond "does the
routine not fail" but we don't want to do too much here).