XSPEC models must now be evaluatedwith lo, hi arrays
Drop support for calling an XSPEC model with a single grid and
have it evaluate it as if called with grid[:-1], grid[1:] and have
a 0 added to the end (to keep the output size to match the input
size). Note that you can still use this approach if you use the
compiled interface to the model (via the _calc method of the
model or directly via functions from sherpa.astro.xspec._xspec).

The reason for this change is primarily to support model
regridding, as we can avoid any special case code for the
XSPEC models if we do not have to support the 'call with only
one array' method. This change however is not added in this PR.

A number of tests of the XSPEC code have been removed because of
this, and a few tweaked to use the lo,hi scheme. This should reduce
the test run-time of builds that include XSPEC support.