Switch from TypeError to FutureWarning
Rather than error out when called with a single grid, create a
FutureWarning, so that users have a chance to update their code.
The issue is whether to use FutureWarning or DeprecationWarning;
I have picked the former as it is shown by default in ipython
sessions, whereas DeprecationWarning is not and so is easy to miss.